Implemented navigation bar
